+++ bundle-es2015/models/Config.d.ts
@@ -2,6 +2,9 @@ import { ResponseHandler, VersionStatus } from 'contensis-core-api';
 export interface Config {
     rootUrl?: string;
     accessToken?: string;
+    defaultHeaders?: {
+        [key: string]: string;
+    };
     projectId?: string;
     language?: string;
     versionStatus?: VersionStatus;