+++ src/models/Config.ts
@@ -3,6 +3,7 @@ import { ResponseHandler, VersionStatus } from 'contensis-core-api';
 export interface Config {
 	rootUrl?: string;
 	accessToken?: string;
+	defaultHeaders?: { [key: string]: string };
 	projectId?: string;
 	language?: string;
 	versionStatus?: VersionStatus;